### 2019-07-11 #1563 @silvandeleemput
> @silvandeleemput if you have any suggestions for potential reviewers then please mention them here in this thread.

Going through the list, I would suggest the following reviewers (having a deep learning background and a preference for Python):

arokem
nilesh-patil
paragkulkarni11
rasbt
sssomani
mamrehn
GregaVrbancic
