### 2019-07-16 #1563 @EduPH
Hi,
It depends on how fast you want an answer. I think I could send it back along first week of August.
