### 2019-07-16 #1576 @whedon
[REVIEW]: MemCNN: a Framework for Developing Memory Efficient Invertible Neural Networks
**Submitting author:** @silvandeleemput (<a href="http://orcid.org/0000-0001-6047-3051">Sil C. van de Leemput</a>)
**Repository:** <a href="https://github.com/silvandeleemput/memcnn" target ="_blank">https://github.com/silvandeleemput/memcnn</a>
**Version:** 0.3.3
