### 2019-07-18 silvandeleemput/memcnn#19 @sepandhaghighi
JOSS Review : Community guidelines
1. Move `.github` folder to root
2. Add `CODE_OF_CONDUCT.md` to `.github` folder
3. Move `CONTRIBUTING.rst` to `.github` folder
4. Add `PULL_REQUEST_TEMPLATE.md` to `.github` folder
