### 2019-07-18 silvandeleemput/memcnn#20 @sepandhaghighi
JOSS Review : Code analysis tools
Analyze your code by some proper tools, these tools helps to identify vulnerability and bugs in source code.

Recommended analyzers :

1. [Codacy](https://www.codacy.com/)
2. [Codefactor](https://www.codefactor.io/)
3. [LGTM](https://lgtm.com/)
4. [Codebeat](https://codebeat.co/)

also take a look at Github Marketplace.
