### 2019-07-18 silvandeleemput/memcnn#21 @sepandhaghighi
JOSS Review : Missing DOIs
Some DOI's missed :

```
OK DOIs

- None

MISSING DOIs

- https://doi.org/10.1109/cvpr.2016.90 may be missing for title: Deep Residual Learning for Image Recognition
- https://doi.org/10.1109/cvprw.2009.5206848 may be missing for title: ImageNet: A Large-Scale Hierarchical Image Database
- https://doi.org/10.1007/978-3-642-35289-8_27 may be missing for title: Training deep and recurrent networks with hessian-free optimization
- https://doi.org/10.1109/iccv.2017.244 may be missing for title: Unpaired image-to-image translation using cycle-consistent adversarial networks

INVALID DOIs

- None
```
