### 2019-07-18 silvandeleemput/memcnn#22 @sepandhaghighi
JOSS Review : Performance & Example usage
The memory consumption table (table 3) is vague, you should compare memory usage between `MemCNN` vs other platforms.

You mentioned :

Table 3 shows the average memory usage during model training using MemCNN. **The results show significant memory savings using the invertible operations as the number of layers increases for ResNet without the use of invertible operations** (Comparison reference ??)
