### 2019-07-18 #1576 @sepandhaghighi
review in progress, would say "accept with minor revisions".
added 4 issues, after those are closed, this should be ready to accept
silvandeleemput/memcnn/issues/22
silvandeleemput/memcnn/issues/21
silvandeleemput/memcnn/issues/20
silvandeleemput/memcnn/issues/19
