### 2019-07-18 fa51777
moved github issue_template to correct place and added github code_of_conduct, contributing, pull_request_template #19
