+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,18 @@
+<!-- Thank you for your contribution to MemCNN! Please replace {Please write here} with your description -->
+
+### What was the problem?
+
+{Please write here}
+
+### How this PR fixes the problem?
+
+{Please write here}
+
+### Check lists (check `x` in `[ ]` of list items)
+
+- [ ] Test passed
+- [ ] Coding style (indentation, etc)
+
+### Additional Comments (if any)
+
+{Please write here}
