### 2019-07-18 a6cd940
Added missing DOIs to paper.bib and corrected formatting #21
