+++ paper/paper.bib
@@ -1,69 +1,70 @@
-
 @article{Gomez17,
-  author    = {A. N. Gomez and
+    author    = {A. N. Gomez and
                M. Ren and
                R. Urtasun and
                R. B. Grosse},
-  title     = {The Reversible Residual Network: Backpropagation Without Storing Activations},
-  journal = {arXiv preprint arXiv:1707.04585},
-  year      = {2017}
+    title     = {The Reversible Residual Network: Backpropagation Without Storing Activations},
+    journal = {arXiv preprint arXiv:1707.04585},
+    year      = {2017}
 }
 
 @article{Dinh14,
-  author    = {L. Dinh and
+    author    = {L. Dinh and
                D. Krueger and
                Y. Bengio},
-  title     = {{NICE:} Non-linear Independent Components Estimation},
-  journal = {arXiv preprint arXiv:1410.8516},
-  year      = {2014}
+    title     = {{NICE:} Non-linear Independent Components Estimation},
+    journal = {arXiv preprint arXiv:1410.8516},
+    year      = {2014}
 }
 
 @article{He2016,
-		author = {K. He and X. Zhang and S. Ren and J. Sun},
-		title = {Identity Mappings in Deep Residual Networks},
-		journal = {arXiv preprint arXiv:1603.05027},
-		year = {2016}
-	}
+    author = {K. He and X. Zhang and S. Ren and J. Sun},
+    title = {Identity Mappings in Deep Residual Networks},
+    journal = {arXiv preprint arXiv:1603.05027},
+    year = {2016},
+    doi = {10.1007/978-3-319-46493-0_38},
+}
 
 @article{He2015,
-		author = {K. He and X. Zhang and S. Ren and J. Sun},
-		title = {Deep Residual Learning for Image Recognition},
-		journal = {arXiv preprint arXiv:1512.03385},
-		year = {2015}
-	}
+    author = {K. He and X. Zhang and S. Ren and J. Sun},
+    title = {Deep Residual Learning for Image Recognition},
+    journal = {arXiv preprint arXiv:1512.03385},
+    year = {2015},
+    doi = {10.1109/cvpr.2016.90}
+}
 
 @article{Chang17,
-  author    = {B. Chang and
+    author    = {B. Chang and
                L. Meng and
                E. Haber and
                L. Ruthotto and
                D. Begert and
                E. Holtham},
-  title     = {Reversible Architectures for Arbitrarily Deep Residual Neural Networks},
-  journal = {arXiv preprint arXiv:1709.03698},
-  year      = {2017}
+    title     = {Reversible Architectures for Arbitrarily Deep Residual Neural Networks},
+    journal = {arXiv preprint arXiv:1709.03698},
+    year      = {2017}
 }
 
 @article{krizhevsky2009learning,
-  title={Learning multiple layers of features from tiny images},
-  author={Krizhevsky, A. and Hinton, G.},
-  year={2009},
-  publisher={Citeseer}
+    title={Learning multiple layers of features from tiny images},
+    author={Krizhevsky, A. and Hinton, G.},
+    year={2009},
+    publisher={Citeseer}
 }
 
 @inproceedings{imagenet_cvpr09,
-        AUTHOR = {Deng, J. and Dong, W. and Socher, R. and Li, L.-J. and Li, K. and Fei-Fei, L.},
-        TITLE = {{ImageNet}: A Large-Scale Hierarchical Image Database},
-        BOOKTITLE = {CVPR09},
-        YEAR = {2009},
-        BIBSOURCE = "http://www.image-net.org/papers/imagenet_cvpr09.bib"
[…]
