### 2019-07-18 silvandeleemput/memcnn#19 @silvandeleemput
Performed 1, 2, and 4 as suggested. For 3, I have added a new `CONTRIBUTING.md` to the `.github` folder with a link which always points to the latest version of the contributing section in the documentation. Moving the `CONTRIBUTING.rst` would result in issues with automated documentation building and packaging.
