### 2019-07-18 silvandeleemput/memcnn#21 @silvandeleemput
The missing DOIs have been added, this issue should be resolved after rebuilding the paper.
