### 2019-07-18 937a147
fixed some minor formatting issues and removed unused code found by codacy #20
