+++ memcnn/models/revop.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import torch
 import torch.nn as nn
 import warnings
 from memcnn.models.additive import AdditiveBlock
