+++ memcnn/utils/tensorboard.py
@@ -10,8 +10,8 @@ def parse_logs(dir_path, output_file = None):
     log_files = glob.glob(os.path.join(dir_path, 'events.out.tfevents.*'))
     event_dict = {}
     # iterate all log_fiiles in order of creation
-    for i, log_file in enumerate(log_files):
-        for j, event in enumerate(tf.train.summary_iterator(log_file)):
+    for log_file in log_files:
+        for event in tf.train.summary_iterator(log_file):
             for value in event.summary.value:
                 if not (value.tag in event_dict):
                     event_dict[value.tag] = collections.OrderedDict()
