### 2019-07-18 214642d
add codacy shield to readme.rst
