### 2019-07-18 6e10703
added .bandit file for ignoring pytests and #nosec comments for known safe asserts #20
