### 2019-07-18 silvandeleemput/memcnn#21 @sepandhaghighi
Thanks for your quick reaction ;-)
Fixed
