+++ .github/CODE_OF_CONDUCT.md
@@ -14,19 +14,19 @@ appearance, race, religion, or sexual identity and orientation.
 Examples of behavior that contributes to creating a positive environment
 include:
 
-  * Using welcoming and inclusive language
-  * Being respectful of differing viewpoints and experiences
-  * Gracefully accepting constructive criticism
-  * Focusing on what is best for the community
-  * Showing empathy towards other community members
+* Using welcoming and inclusive language
+* Being respectful of differing viewpoints and experiences
+* Gracefully accepting constructive criticism
+* Focusing on what is best for the community
+* Showing empathy towards other community members
 
 Examples of unacceptable behavior by participants include:
 
-  * The use of sexualized language or imagery and unwelcome sexual attention or advances
-  * Trolling, insulting/derogatory comments, and personal or political attacks
-  * Public or private harassment
-  * Publishing others' private information, such as a physical or electronic address, without explicit permission
-  * Other conduct which could reasonably be considered inappropriate in a professional setting
+* The use of sexualized language or imagery and unwelcome sexual attention or advances
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information, such as a physical or electronic address, without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a professional setting
 
 ## Our Responsibilities
 
