### 2019-07-19 3bd41b0
Reworked tables on memory usage and the explanations in paper.md and readme.rst #22
