### 2019-07-19 silvandeleemput/memcnn#20 @sepandhaghighi
> For this issue, I have configured the repository for automatic code quality analysis by Codacy.
> I have also added the code quality Codacy badge to the readme.rst.
>
> Some remarks here:
>
> * bandit's asserts security vulnerabilities were disabled since I use those a lot for pytests and none of them were actually security concerns.
> * paper.md was excluded since it has a lot of minor formatting issues because of the integration with LaTeX.
>
> @sepandhaghighi Would this be sufficient for closing this issue?

Yeah, sufficient for this point.
