### 2019-07-19 silvandeleemput/memcnn#19 @sepandhaghighi
> @sepandhaghighi I think this can be closed. Do you agree?

Yes, fixed ;-)
