### 2019-07-19 silvandeleemput/memcnn#22 @silvandeleemput
@sepandhaghighi Hi, I agree with you that the table description and purpose were not very clear. The intent of this table was to compare similar CNN implementations (all in PyTorch) with (RevNet) and without (ResNet) memory savings against each other. Here the RevNet implementations use the MemCNN reversible block. So the comparisons are: ResNet-32 vs RevNet-38, ResNet-110 vs RevNet-110, and ResNet-164 vs RevNet-164. The comparison shows that the use of MemCNN (RevNet implementations) achieves significant memory savings for the activations over conventional implementations (ResNet implementations).

Additionally, I found the table lacking some important details so I have split the memory usage into parameter memory usage and activations memory usage, which required recomputing the memory usage altogether. Furthermore, I have added the number of parameters for each model to the table as well.

The new statistics have been added to the table and the description has been changed to clarify the table's purpose. These changes affect both the `paper.md` and the `readme.rst`.
