### 2019-07-22 silvandeleemput/memcnn#22 @sepandhaghighi
@silvandeleemput Thanks for your efforts, now it's meaningful.
As a final suggestion, I think it will better if you show related items in one row (for example `ResNet-3`2 vs `RevNet-38`) or show `RevNet` and `ResNet` details in separate tables (peer to peer).

 Notice  : For both `Table3` and `Table2`
