### 2019-07-22 silvandeleemput/memcnn#23 @sepandhaghighi
JOSS Review : Add memcnn to Anaconda Cloud
Most of machine learning libraries  are available in `conda` package manager.
You can add your package directly from `PyPI` to `Anaconda Cloud` by using of `skeleton` and `conda-build`.
