### 2019-07-22 #1576 @sepandhaghighi
another suggestion ;-)
silvandeleemput/memcnn/issues/23
