### 2019-07-23 silvandeleemput/memcnn#23 @silvandeleemput
@sepandhaghighi Hi, thanks for your suggestion. I really like this idea. I have tried to do as you suggested and executed the `skeleton` command to generate a build from `PyPI` directly. However, it appears I am missing an include statement in my `manifest.in`.  This implies I would need to first roll out a new version release to PyPI before I can use this approach.

It seems to me that your suggested approach is the most simple/desirable, so I would like to stick with it if possible. However, does changing the package version (probably 0.3.3 -> 0.3.4) not conflict with the JOSS review?
