### 2019-07-23 46f24a8
added requirements.txt includes to manifest.in #23
