### 2019-07-23 silvandeleemput/memcnn#23 @sepandhaghighi
@silvandeleemput
Hi
You can close this issue at the end.
Wait until the end of the review process, then you should release a new version that include all edits and provide an associated DOI (e.g via `Zenodo`).
