### 2019-07-23 74c769a
further clarified result tables in readme.rst and paper.md #22
