+++ paper/paper.md
@@ -131,44 +131,62 @@ The reversible block model has an advantageous memory footprint when chained in 
 
 # Experiments and results
 
- **Table 2:** Accuracy (acc.) and training time (time, in hours:minutes) comparison of the PyTorch implementation (MemCNN) versus the Tensorflow implementation from [@Gomez17] on Cifar-10 and Cifar-100 [@krizhevsky2009learning].
+ **Table 2a:** Accuracy comparison of the PyTorch implementation (MemCNN) versus the Tensorflow implementation from [@Gomez17] on Cifar-10 and Cifar-100 [@krizhevsky2009learning]. Accuracies were approximately similar between implementations.
  
 \begin{center}
  
 \vspace{0.3cm}
  
-\begin{tabular}{lcccccccc}
-\hline         & \multicolumn{4}{c}{\textbf{Tensorflow}} & \multicolumn{4}{c}{\textbf{PyTorch}} \\
-           & \multicolumn{2}{c}{\textbf{Cifar-10}} & \multicolumn{2}{c}{\textbf{Cifar-100}} & \multicolumn{2}{c}{\textbf{Cifar-10}} & \multicolumn{2}{c}{\textbf{Cifar-100}} \\
-\textbf{Model} & \textbf{acc.} & \textbf{time} & \textbf{acc.} & \textbf{time} & \textbf{acc.} & \textbf{time} & \textbf{acc.} & \textbf{time} \\\hline
-resnet-32  & 92.74  & \,\,2:04  & 69.10   & \,\,1:58   & 92.86  & 1:51    & 69.81   & 1:51    \\
-resnet-110 & 93.99  & \,\,4:11 & 73.30   & \,\,6:44   & 93.55  & 2:51    & 72.40   & 2:39    \\
-resnet-164 & 94.57  & 11:05 & 76.79   & 10:59  & 94.80  & 4:59    & 76.47   & 3:45    \\
-revnet-38  & 93.14  & \,\,2:17 & 71.17   & \,\,2:20   & 92.80  & 2:09    & 69.90   & 2:16    \\
-revnet-110 & 94.02  & \,\,6:59 & 74.00   & \,\,7:03    & 94.10  & 3:42    & 73.30   & 3:50    \\
-revnet-164 & 94.56  & 13:09 & 76.39   & 13:12  & 94.90  & 7:21    & 76.90   & 7:17    \\ \hline
+\begin{tabular}{lcccc}
+\hline         & \multicolumn{2}{c}{\textbf{Cifar-10}} & \multicolumn{2}{c}{\textbf{Cifar-100}} \\
+\textbf{Model} & \textbf{Tensorflow} & \textbf{PyTorch} & \textbf{Tensorflow} & \textbf{PyTorch} \\ \hline
+ResNet-32  & 92.74  & 92.86  & 69.10  & 69.81 \\
+ResNet-110 & 93.99  & 93.55  & 73.30  & 72.40 \\
+ResNet-164 & 94.57  & 94.80  & 76.79  & 76.47 \\
+RevNet-38  & 93.14  & 92.80  & 71.17  & 69.90 \\
+RevNet-110 & 94.02  & 94.10  & 74.00  & 73.30 \\
+RevNet-164 & 94.56  & 94.90  & 76.39  & 76.90 \\ \hline
+\end{tabular}
+
+\vspace{0.3cm}
+
+\end{center}
+
[…]
