+++ README.rst
@@ -118,44 +118,66 @@ TensorFlow results were obtained from `the reversible residual
 network <https://arxiv.org/abs/1707.04585>`__ running the code from
 their `GitHub <https://github.com/renmengye/revnet-public>`__.
 
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-|            |               Tensorflow                          |               PyTorch                       |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-|            |    Cifar-10            |      Cifar-100           |      Cifar-10        |     Cifar-100        |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| Model      | acc.     | time        | acc.      | time         | acc.     | time      | acc.      | time     |
-+============+==========+=============+===========+==============+==========+===========+===========+==========+
-| resnet-32  |  92.74   |  2:04       |  69.10    |      1:58    |  92.86   |  1:51     |  69.81    |  1:51    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| resnet-110 |  93.99   |  4:11       |  73.30    |      6:44    |  93.55   |  2:51     |  72.40    |  2:39    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| resnet-164 |  94.57   | 11:05       |  76.79    |  10:59       |  94.80   |  4:59     |  76.47    |  3:45    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| revnet-38  |  93.14   |  2:17       |  71.17    |      2:20    |  92.80   |  2:09     |  69.90    |  2:16    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| revnet-110 |  94.02   |  6:59       |  74.00    |      7:03    |  94.10   |  3:42     |  73.30    |  3:50    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-| revnet-164 |  94.56   | 13:09       |  76.39    |  13:12       |  94.90   |  7:21     |  76.90    |  7:17    |
-+------------+----------+-------------+-----------+--------------+----------+-----------+-----------+----------+
-
-
 The PyTorch results listed were recomputed on June 11th 2018, and differ
 from the results in the ICLR paper. The Tensorflow results are still the
 same.
 
+Prediction accuracy
+^^^^^^^^^^^^^^^^^^^
+
++------------+------------------------+--------------------------+----------------------+----------------------+
+|            |               Cifar-10                            |               Cifar-100                     |
++------------+------------------------+--------------------------+----------------------+----------------------+
+| Model      |    Tensorflow          |      PyTorch             |      Tensorflow      |     PyTorch          |
++============+========================+==========================+======================+======================+
+| resnet-32  |  92.74                 |    92.86                 |   69.10              |  69.81               |
++------------+------------------------+--------------------------+----------------------+----------------------+
+| resnet-110 |  93.99                 |    93.55                 |   73.30              |  72.40               |
++------------+------------------------+--------------------------+----------------------+----------------------+
+| resnet-164 |  94.57                 |    94.80                 |   76.79              |  76.47               |
++------------+------------------------+--------------------------+----------------------+----------------------+
+| revnet-38  |  93.14                 |    92.80                 |   71.17              |  69.90               |
++------------+------------------------+--------------------------+----------------------+----------------------+
+| revnet-110 |  94.02                 |    94.10                 |   74.00              |  73.30               |
++------------+------------------------+--------------------------+----------------------+----------------------+
[…]
