### 2019-07-23 silvandeleemput/memcnn#22 @silvandeleemput
@sepandhaghighi Good suggestion! I have changed `Table 3` to have a row-wise comparison for `Resnet` vs `RevNet`. Furthermore, I have split `Table 2` into `Table 2a` and `Table 2b` for prediction accuracy and training time respectively and reordered the headers such that the comparison between the Tensorflow versus PyTorch implementations is easier to perform. Overall, these changes should facilitate the intended comparisons.

Changes have been made to both the `paper.md` and the `readme.rst`.
