### 2019-07-24 silvandeleemput/memcnn#22 @sepandhaghighi
Well done!
