### 2019-07-25 silvandeleemput/memcnn#24 @GregaVrbancic
JOSS Review: Run Experiments
After installation of memcnn into the fresh python environment, I have followed the [documentation](https://memcnn.readthedocs.io/en/latest/usage.html#run-pytorch-experiments) on how to run the experiments. For me, it was a bit unclear what was the problem with the config.json file, when the first time running the experiment  I experienced the error of config.json not existing error, as in the instructions there is no mention of it. I have fixed this by copying the config.json.example in the required location.

@silvandeleemput  please take a look at this issue and try to address this with improving the documentation on how to run experiments. After improving the documentation you can close this issue and check back in the JOSS review issue.
