### 2019-07-25 #1576 @GregaVrbancic
Hi @silvandeleemput @terrytangyuan, sorry for a bit of a delay, I was swamped at work. I finally had time to go through the remaining items from the checklist. In addition to the already addressed remarks from @sepandhaghighi, @silvandeleemput please take a look at this issuehttps://github.com/silvandeleemput/memcnn/issues/24
