### 2019-07-26 4433657
Added documentation on the experiment configuration file to README.rst, installation.rst, and train.py #24
