### 2019-07-26 silvandeleemput/memcnn#24 @silvandeleemput
@GregaVrbancic Thank you for identifying this unclarity. I assume you ran into this issue after installing the package using `pip install` in your environment, which does not automatically set the configuration file (in contrast to the docker images, which do this automatically). I changed the documentation (`readme.rst`, `installation.rst`) to mention the necessity to set the configuration file manually and how to configure it. As a bonus, I also changed `train.py` to mention the configuration file when it is missing or improperly configured, to help the user resolve the issue faster.

Do these changes suffice for closing this issue?
