### 2019-07-26 silvandeleemput/memcnn#24 @GregaVrbancic
@silvandeleemput great, nice work! This issue can be closed now. I'll notify the editor that my review is done, so the publishing process can proceed.
