### 2019-07-26 silvandeleemput/memcnn#24 @silvandeleemput
@GregaVrbancic thanks for your quick response! I'll close the issue.
