### 2019-07-26 #1576 @GregaVrbancic
@terrytangyuan I'm satisfied with the changes made. It looks good to me for publishing.
