+++ memcnn/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = """Sil van de Leemput"""
 __email__ = 'silvandeleemput@gmail.com'
-__version__ = '0.3.3'
+__version__ = '0.3.4'
 
 
 from memcnn.models.revop import ReversibleBlock
