### 2019-07-28 0d8feea
Anaconda cloud circleci integration, documentation for installation, new badge and version update #23
