### 2019-07-28 12d2fa2
minor change to README.rst
