### 2019-07-28 bf26a31
updated version history and minor timeout tweaks to circleci config.yml
