### 2019-07-28 f856b7f
Bump version: 0.3.5 → 1.0.0
