+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.5
+current_version = 1.0.0
 commit = True
 tag = True
 tag_name = {new_version}
