### 2019-07-28 #1576 @silvandeleemput
@sepandhaghighi @GregaVrbancic Thank you both for your time and effort in reviewing this work.

@terrytangyuan I have created a new release of MemCNN (1.0.0) with the changes from the review included and I have uploaded it to Zenodo and saved it as an unpublished upload. The DOI of the archive is: `10.5281/zenodo.3353604`. The title and the authors of the submission are the same as in the paper and should match your requirements.
