### 2019-07-28 caa972a
minor installation documentation corrections
