### 2019-07-28 silvandeleemput/memcnn#23 @silvandeleemput
I have just released a new version of MemCNN (1.0.0) which contains all the changes from the review and which has a functional Anaconda cloud integration. See: https://anaconda.org/silvandeleemput/memcnn
Also, I have added an anaconda badge in the `readme.rst` and installation instructions for this as well.

I think this issue can be closed as well.
