### 2019-07-28 #1576 @terrytangyuan
@openjournals/joss-eics Since this is the first accepted paper I am editing, could you help double check? Also, it looks like `@whedon accept deposit=true` can only be done by editors-in-chief so I am handing this over to you now.
