### 2019-07-29 #1576 @kyleniemeyer
@terrytangyuan sure thing. This is the right point to tag us; I do the following:

- Double check that all the review items have been checked off and issues resolved.
- For us to publish, we should see that the archive DOI resolves and that the metadata is correct; @silvandeleemput, can you publish the latest version? We are done with any code changes.
- I next go through the paper to look for any typos, errors in references/citations, or anything else that stands out. Another thing I typically notice is incomplete affiliation info (should generally be something like department, institution, city, country—what you'd see in any journal article).

Also, an explicit `@whedon check references` check at the end is helpful, but is not a perfect tool.

The AEiCs will always do these checks when you hand it off to us, but it's certainly helpful for you to take care of any such issues before it gets to us.
