### 2019-07-29 #1576 @silvandeleemput
@kyleniemeyer @terrytangyuan The Zenodo code archive is now published (MemCNN version 1.0.0) and can be found here: https://zenodo.org/record/3353604#.XT6VqOgzaUk
