+++ paper/paper.bib
@@ -1,39 +1,47 @@
-@article{Gomez17,
+@misc{Gomez17,
     author    = {A. N. Gomez and
                M. Ren and
                R. Urtasun and
                R. B. Grosse},
     title     = {The Reversible Residual Network: Backpropagation Without Storing Activations},
-    journal = {arXiv preprint arXiv:1707.04585},
-    year      = {2017}
+    howpublished = {{\tt arXiv:1707.04585 [cs.CV]}},
+    url = {https://arxiv.org/abs/1707.04585},
+    year      = 2017
 }
 
-@article{Dinh14,
+@misc{Dinh14,
     author    = {L. Dinh and
                D. Krueger and
                Y. Bengio},
     title     = {{NICE:} Non-linear Independent Components Estimation},
-    journal = {arXiv preprint arXiv:1410.8516},
-    year      = {2014}
+    howpublished = {{\tt arXiv:1410.8516 [cs.LG]}},
+    url = {https://arxiv.org/abs/1410.8516},
+    year      = 2014
 }
 
 @article{He2016,
-    author = {K. He and X. Zhang and S. Ren and J. Sun},
     title = {Identity Mappings in Deep Residual Networks},
-    journal = {arXiv preprint arXiv:1603.05027},
-    year = {2016},
+    note = {{\tt arXiv:1603.05027 [cs.CV]}},
+    year = 2016,
     doi = {10.1007/978-3-319-46493-0_38},
+  publisher = {Springer International Publishing},
+  pages = {630--645},
+  author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
+  booktitle = {Computer Vision {\textendash} {ECCV} 2016}
 }
 
 @article{He2015,
-    author = {K. He and X. Zhang and S. Ren and J. Sun},
-    title = {Deep Residual Learning for Image Recognition},
-    journal = {arXiv preprint arXiv:1512.03385},
-    year = {2015},
-    doi = {10.1109/cvpr.2016.90}
+  note = {{\tt arXiv:1512.03385 [cs.CV]}},
+  doi = {10.1109/cvpr.2016.90},
+  year = {2016},
+  month = jun,
+  publisher = {{IEEE}},
+  author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
+  title = {Deep Residual Learning for Image Recognition},
+  booktitle = {2016 {IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})}
 }
 
-@article{Chang17,
+@misc{Chang17,
     author    = {B. Chang and
                L. Meng and
                E. Haber and
@@ -41,34 +49,40 @@
                D. Begert and
                E. Holtham},
     title     = {Reversible Architectures for Arbitrarily Deep Residual Neural Networks},
-    journal = {arXiv preprint arXiv:1709.03698},
-    year      = {2017}
+    howpublished = {{\tt arXiv:1709.03698 [cs.CV]}},
+    url = {https://arxiv.org/abs/1709.03698},
+    year      = 2017
 }
 
-@article{krizhevsky2009learning,
-    title={Learning multiple layers of features from tiny images},
-    author={Krizhevsky, A. and Hinton, G.},
-    year={2009},
-    publisher={Citeseer}
+@mastersthesis{krizhevsky2009learning,
+  author = {Krizhevsky, A.},
+  title = {Learning Multiple Layers of Features from Tiny Images},
+  school       = {University of Toronto},
+  year         = 2009,
+  address      = {Toronto, Ontario, Canada},
+  month        = apr,
 }
 
 @inproceedings{imagenet_cvpr09,
[…]
