### 2019-07-29 silvandeleemput/memcnn#23 @sepandhaghighi
👍👍 fixed ;-)
