### 2019-07-29 #1576 @kyleniemeyer
Thanks @silvandeleemput; can you merge this PR that corrects some references? https://github.com/silvandeleemput/memcnn/pull/25

I'll be looking through the paper itself next.
