### 2019-07-29 #1576 @silvandeleemput
@kyleniemeyer Thanks! I have merged the PR.
