### 2019-07-29 #1576 @kyleniemeyer
Hi @silvandeleemput, for the paper, I've made some edits to in-text citations in https://github.com/silvandeleemput/memcnn/pull/26.

My only other requested change is for you to add, from our list of "[what should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)":

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

Thanks!
