### 2019-07-29 #1576 @silvandeleemput
@kyleniemeyer We are unsure what exactly is meant with a "non-specialist audience" since it is very likely that people finding this work are at least familiar with neural networks, Python, and/or PyTorch, which should be sufficient to understand the purpose of the library from the paper.

If we are to add the requested change, what section of the paper do you recommend for this?
