### 2019-07-29 #1576 @kyleniemeyer
@silvandeleemput We just ask that all papers try to have a sentence or two at the beginning explaining the purpose in a way such that a reader from a non-expert background could understand (in a shallow way) what the application is.

In your case, you might add one or two sentences at the very beginning explaining classification problems and/or neural networks for a general reader.
