### 2019-07-30 0325cdc
added a high level explanation about neural networks to the summary and updated some bib items
