+++ paper/paper.bib
@@ -24,21 +24,21 @@
     note = {{\tt arXiv:1603.05027 [cs.CV]}},
     year = 2016,
     doi = {10.1007/978-3-319-46493-0_38},
-  publisher = {Springer International Publishing},
-  pages = {630--645},
-  author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
-  booktitle = {Computer Vision {\textendash} {ECCV} 2016}
+    publisher = {Springer International Publishing},
+    pages = {630--645},
+    author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
+    booktitle = {Computer Vision {\textendash} {ECCV} 2016}
 }
 
 @article{He2015,
-  note = {{\tt arXiv:1512.03385 [cs.CV]}},
-  doi = {10.1109/cvpr.2016.90},
-  year = {2016},
-  month = jun,
-  publisher = {{IEEE}},
-  author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
-  title = {Deep Residual Learning for Image Recognition},
-  booktitle = {2016 {IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})}
+    note = {{\tt arXiv:1512.03385 [cs.CV]}},
+    doi = {10.1109/cvpr.2016.90},
+    year = {2016},
+    month = jun,
+    publisher = {{IEEE}},
+    author = {Kaiming He and Xiangyu Zhang and Shaoqing Ren and Jian Sun},
+    title = {Deep Residual Learning for Image Recognition},
+    booktitle = {2016 {IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})}
 }
 
 @misc{Chang17,
@@ -55,22 +55,22 @@
 }
 
 @mastersthesis{krizhevsky2009learning,
-  author = {Krizhevsky, A.},
-  title = {Learning Multiple Layers of Features from Tiny Images},
-  school       = {University of Toronto},
-  year         = 2009,
-  address      = {Toronto, Ontario, Canada},
-  month        = apr,
+    author = {Krizhevsky, A.},
+    title = {Learning Multiple Layers of Features from Tiny Images},
+    school       = {University of Toronto},
+    year         = 2009,
+    address      = {Toronto, Ontario, Canada},
+    month        = apr,
 }
 
 @inproceedings{imagenet_cvpr09,
     doi = {10.1109/cvprw.2009.5206848},
     year = 2009,
-  month = jun,
-  publisher = {{IEEE}},
-  author = {Jia Deng and Wei Dong and Richard Socher and Li-Jia Li and  Kai Li and  Li Fei-Fei},
-  title = {{ImageNet}: A large-scale hierarchical image database},
-  booktitle = {2009 {IEEE} Conference on Computer Vision and Pattern Recognition}
+    month = jun,
+    publisher = {{IEEE}},
+    author = {Jia Deng and Wei Dong and Richard Socher and Li-Jia Li and  Kai Li and  Li Fei-Fei},
+    title = {{ImageNet}: A large-scale hierarchical image database},
+    booktitle = {2009 {IEEE} Conference on Computer Vision and Pattern Recognition}
 }
 
 @inproceedings{jaco18,
@@ -78,53 +78,54 @@
     title = {{i-RevNet}: Deep Invertible Networks},
     booktitle = {ICLR},
     year = {2018},
+    url = {https://arxiv.org/abs/1802.07088},
     howpublished = {{\tt arXiv:1802.07088 [cs.LG]}}
 }
 
 @misc{TF2015,
[…]
