### 2019-07-30 #1576 @silvandeleemput
@kyleniemeyer Thank you for clarifying your recommendation. We have added a high-level explanation about neural networks to the summary. Additionally, we have also added a missing arxiv URL to the bibliography so that all works have either an URL or DOI.
