+++ paper/paper.bib
@@ -181,7 +181,7 @@
     booktitle = {{The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)}},
     month = jun,
     year = 2019,
-    howpublished = {{\tt arXiv:1902.02729 [cs.LG]}},
+    howpublished = {{\tt arXiv:1902.02729 [cs.CV]}},
     url = {https://arxiv.org/abs/1902.02729},
 }
 
