### 2019-07-30 #1576 @kyleniemeyer
Congrats @silvandeleemput on your submission's publication in JOSS! Thanks to @GregaVrbancic and @sepandhaghighi for reviewing, and @terrytangyuan for editing!
