### 2019-07-30 #1576 @silvandeleemput
@kyleniemeyer @terrytangyuan Hi thank you both very much for accepting our work and for helping us to publish our paper!

There is just a minor issue regarding the citation information on the [linked DOI url](http://joss.theoj.org/papers/10.21105/joss.01576): my surname is missing the "C." and has the "van de" in camel case while it should be all in lowercase  (as it is written in the paper). Similarly, for the third author "Bram van Ginneken" the "van" should be all in lowercase. Can this be corrected?

See the corrected bib file below:

```bib
@article{vandeLeemput2019MemCNN,
  journal = {Journal of Open Source Software},
  doi = {10.21105/joss.01576},
  issn = {2475-9066},
  number = {39},
  publisher = {The Open Journal},
  title = {MemCNN: A Python/PyTorch package for creating memory-efficient invertible neural networks},
  url = {http://dx.doi.org/10.21105/joss.01576},
  volume = {4},
  author = {Sil C. {van de} Leemput and Jonas Teuwen and Bram {van} Ginneken and Rashindra Manniesing},
  pages = {1576},
  date = {2019-07-30},
  year = {2019},
  month = {7},
  day = {30},
}

```
