### 2019-07-31 #1576 @arfon
> @arfon can you make this correction manually? Also, any idea about how to avoid such issues in the future? This isn't the first surname issue we've had recently.

Done. I'm afraid the name parsing library we use doesn't handle surnames names like this very well so fixing is non-trivial.
